mitosheet: polish up the frontend
